#72 Update to contollers, models, traits
