+++ app/Concerns/HasTeam.php
@@ -7,9 +7,6 @@ use Illuminate\Database\Eloquent\Builder;
 /**
  * Trait for managing team context on User models.
  *
- * This trait requires the model to use Spatie\Permission\Traits\HasRoles
- * for the forgetCachedPermissions() method.
- *
  * @mixin \Spatie\Permission\Traits\HasRoles
  *
  * @property int|null $team_id
@@ -60,8 +57,6 @@ trait HasTeam
     {
         $this->team_id = $teamId;
 
-        // Call parent trait method - provided by
-        // Spatie\Permission\Traits\HasRoles
         /** @var \Spatie\Permission\Traits\HasRoles $this */
         $this->forgetCachedPermissions();
