+++ app/Concerns/PasswordValidationRules.php
@@ -10,7 +10,7 @@ trait PasswordValidationRules
     /**
      * Get the validation rules used to validate passwords.
      *
-     * @return array<int, ValidationRule|array<mixed>|string>
+     * @return array<int,ValidationRule|array<mixed>|string>
      */
     protected function passwordRules(): array
     {
@@ -20,7 +20,7 @@ trait PasswordValidationRules
     /**
      * Get the validation rules used to validate the current password.
      *
-     * @return array<int, ValidationRule|array<mixed>|string>
+     * @return array<int,ValidationRule|array<mixed>|string>
      */
     protected function currentPasswordRules(): array
     {
