+++ app/Concerns/ProfileValidationRules.php
@@ -11,7 +11,7 @@ trait ProfileValidationRules
     /**
      * Get the validation rules used to validate user profiles.
      *
-     * @return array<string, array<int, ValidationRule|array<mixed>|string>>
+     * @return array<string, array<int,ValidationRule|array<mixed>|string>>
      */
     protected function profileRules(?int $userId = null): array
     {
@@ -24,7 +24,7 @@ trait ProfileValidationRules
     /**
      * Get the validation rules used to validate user names.
      *
-     * @return array<int, ValidationRule|array<mixed>|string>
+     * @return array<int,ValidationRule|array<mixed>|string>
      */
     protected function nameRules(): array
     {
@@ -34,7 +34,7 @@ trait ProfileValidationRules
     /**
      * Get the validation rules used to validate user emails.
      *
-     * @return array<int, ValidationRule|array<mixed>|string>
+     * @return array<int,ValidationRule|array<mixed>|string>
      */
     protected function emailRules(?int $userId = null): array
     {
