+++ app/Concerns/Users/HasUserRoles.php
@@ -8,35 +8,27 @@ use Spatie\Permission\Contracts\Role;
 /**
  * Trait for managing specialized user roles beyond base admin/user roles.
  *
- * This trait requires the model to use Spatie\Permission\Traits\HasRoles.
- * The assignRole() and removeRole() methods are provided by that trait.
- *
  * @mixin \Spatie\Permission\Traits\HasRoles
  *
- * @property \Illuminate\Support\Collection $roles
- * @property \Illuminate\Support\Collection $permissions
- * @property \Illuminate\Support\Collection $teams
- * @property \Illuminate\Support\Collection $specialisedRoles
+ * @property Collection $roles
+ * @property Collection $permissions
+ * @property Collection $teams
+ * @property Collection $specialisedRoles
  *
- * @property-read \Illuminate\Support\Collection $specialisedRoles
+ * @property-read Collection $specialisedRoles
  */
 trait HasUserRoles
 {
     /**
      * Assign one or more specialised roles to the user.
      *
-     * This method wraps the HasRoles::assignRole() method from Spatie.
-     *
-     * @param  string|array|Role|\Illuminate\Support\Collection $roles
-     * Role name(s) to assign.
+     * @param  string|array|Role|Collection $roles
      *
      * @return self
      */
     public function assignSpecialisedRole(
         string|array|Role|Collection $roles
     ): self {
-        // Call parent trait method - provided by
-        // Spatie\Permission\Traits\HasRoles
         /** @var \Spatie\Permission\Traits\HasRoles $this */
         $this->assignRole($roles);
         return $this;
@@ -45,16 +37,12 @@ trait HasUserRoles
     /**
      * Remove one or more specialised roles from the user.
      *
-     * This method wraps the HasRoles::removeRole() method from Spatie.
-     *
-     * @param  string|array|Role $roles Role name(s) to remove.
+     * @param  string|array|Role $roles
      *
      * @return self
      */
     public function removeSpecialisedRole(string|array|Role $roles): self
     {
-        // Call parent trait method -
-        // provided by Spatie\Permission\Traits\HasRoles
         /** @var \Spatie\Permission\Traits\HasRoles $this */
         $this->removeRole($roles);
         return $this;
