+++ app/Concerns/Users/HasUserScopes.php
@@ -4,11 +4,6 @@ namespace App\Concerns\Users;
 
 use Illuminate\Database\Eloquent\Builder;
 
-/**
- * Trait providing query scopes for User model filtering.
- *
- * @mixin \App\Models\User
- */
 trait HasUserScopes
 {
     /**
